# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shop Rental Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type AuthSessionUser = {
  id: string;
  name: string;
  username: string;
  role: string;
};

export interface AuthSession {
  authenticated: boolean;
  user?: AuthSessionUser;
}

export interface Tenant {
  id: string;
  name: string;
  phone: string;
  nidNumber?: string;
  address?: string;
  email?: string;
  notes?: string;
  createdAt: string;
}

export interface TenantInput {
  name: string;
  phone: string;
  nidNumber?: string;
  address?: string;
  email?: string;
  notes?: string;
}

export type ShopStatus = (typeof ShopStatus)[keyof typeof ShopStatus];

export const ShopStatus = {
  vacant: "vacant",
  occupied: "occupied",
} as const;

export interface Shop {
  id: string;
  /** Shop number / code */
  code: string;
  /** Market / floor / area */
  location: string;
  sizeSqft?: number;
  monthlyRent: number;
  depositAmount?: number;
  status: ShopStatus;
  tenantId?: string;
  tenantName?: string;
  leaseStart?: string;
  notes?: string;
  createdAt: string;
}

export interface ShopInput {
  code: string;
  location: string;
  sizeSqft?: number;
  monthlyRent: number;
  depositAmount?: number;
  notes?: string;
}

export interface AssignTenantInput {
  /** Pass null to unassign */
  tenantId?: string | null;
  leaseStart?: string;
}

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  paid: "paid",
  unpaid: "unpaid",
  partial: "partial",
} as const;

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod];

export const PaymentMethod = {
  cash: "cash",
  bkash: "bkash",
  nagad: "nagad",
  rocket: "rocket",
  bank: "bank",
  other: "other",
} as const;

export interface Payment {
  id: string;
  shopId: string;
  shopCode: string;
  tenantId: string;
  tenantName: string;
  /** YYYY-MM */
  month: string;
  amountDue: number;
  amountPaid: number;
  status: PaymentStatus;
  paidOn?: string;
  method?: PaymentMethod;
  note?: string;
  createdAt: string;
}

export type PaymentInputMethod =
  (typeof PaymentInputMethod)[keyof typeof PaymentInputMethod];

export const PaymentInputMethod = {
  cash: "cash",
  bkash: "bkash",
  nagad: "nagad",
  rocket: "rocket",
  bank: "bank",
  other: "other",
} as const;

export interface PaymentInput {
  shopId: string;
  month: string;
  amountDue: number;
  amountPaid: number;
  paidOn?: string;
  method?: PaymentInputMethod;
  note?: string;
}

export interface GenerateMonthInput {
  /** YYYY-MM */
  month: string;
}

export interface DashboardSummary {
  totalShops: number;
  occupiedShops: number;
  vacantShops: number;
  totalTenants: number;
  currentMonth: string;
  expectedThisMonth: number;
  collectedThisMonth: number;
  outstandingThisMonth: number;
  paidCount: number;
  unpaidCount: number;
}

export interface DueReminder {
  paymentId: string;
  shopCode: string;
  tenantName: string;
  tenantPhone: string;
  month: string;
  amountDue: number;
  amountPaid: number;
  daysOverdue: number;
}

export interface MonthlyCollectionPoint {
  month: string;
  expected: number;
  collected: number;
}

export interface OccupancySummary {
  occupied: number;
  vacant: number;
}

export interface TopTenant {
  tenantId: string;
  tenantName: string;
  totalPaid: number;
}

export type ListPaymentsParams = {
  /**
   * YYYY-MM
   */
  month?: string;
  status?: ListPaymentsStatus;
  tenantId?: string;
  shopId?: string;
};

export type ListPaymentsStatus =
  (typeof ListPaymentsStatus)[keyof typeof ListPaymentsStatus];

export const ListPaymentsStatus = {
  paid: "paid",
  unpaid: "unpaid",
  partial: "partial",
} as const;