# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shop Rental Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: tenants
  - name: shops
  - name: payments
  - name: dashboard
  - name: reports
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Landlord/admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"

  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    put:
      operationId: updateTenant
      tags: [tenants]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /shops:
    get:
      operationId: listShops
      tags: [shops]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"
    post:
      operationId: createShop
      tags: [shops]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /shops/{id}:
    get:
      operationId: getShop
      tags: [shops]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    put:
      operationId: updateShop
      tags: [shops]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    delete:
      operationId: deleteShop
      tags: [shops]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /shops/{id}/assign:
    post:
      operationId: assignTenantToShop
      tags: [shops]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTenantInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"

  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      parameters:
        - in: query
          name: month
          schema:
            type: string
            description: YYYY-MM
        - in: query
          name: status
          schema:
            type: string
            enum: [paid, unpaid, partial]
        - in: query
          name: tenantId
          schema:
            type: string
        - in: query
          name: shopId
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    put:
      operationId: updatePayment
      tags: [payments]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /payments/generate:
    post:
      operationId: generateMonthlyPayments
      tags: [payments]
      summary: Generate unpaid invoices for a month for all assigned shops
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMonthInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: KPIs for the landlord dashboard
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/recent-payments:
    get:
      operationId: getRecentPayments
      tags: [dashboard]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /dashboard/due-reminders:
    get:
      operationId: getDueReminders
      tags: [dashboard]
      summary: Outstanding invoices that need follow-up
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DueReminder"

  /reports/monthly-collection:
    get:
      operationId: getMonthlyCollection
      tags: [reports]
      summary: Last 12 months of collection vs expected
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyCollectionPoint"
  /reports/shop-occupancy:
    get:
      operationId: getShopOccupancy
      tags: [reports]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OccupancySummary"
  /reports/top-tenants:
    get:
      operationId: getTopTenants
      tags: [reports]
      summary: Tenants ranked by total paid this year
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopTenant"

components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status: { type: string }
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AuthSession:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }
        user:
          type: object
          required: [id, name, username, role]
          properties:
            id: { type: string }
            name: { type: string }
            username: { type: string }
            role: { type: string }

    Tenant:
      type: object
      required: [id, name, phone, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        nidNumber: { type: string }
        address: { type: string }
        email: { type: string }
        notes: { type: string }
        createdAt: { type: string, format: date-time }
    TenantInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string }
        phone: { type: string }
        nidNumber: { type: string }
        address: { type: string }
        email: { type: string }
        notes: { type: string }

    Shop:
      type: object
      required: [id, code, location, monthlyRent, status, createdAt]
      properties:
        id: { type: string }
        code: { type: string, description: "Shop number / code" }
        location: { type: string, description: "Market / floor / area" }
        sizeSqft: { type: number }
        monthlyRent: { type: number }
        depositAmount: { type: number }
        status:
          type: string
          enum: [vacant, occupied]
        tenantId: { type: string }
        tenantName: { type: string }
        leaseStart: { type: string, format: date }
        notes: { type: string }
        createdAt: { type: string, format: date-time }
    ShopInput:
      type: object
      required: [code, location, monthlyRent]
      properties:
        code: { type: string }
        location: { type: string }
        sizeSqft: { type: number }
        monthlyRent: { type: number }
        depositAmount: { type: number }
        notes: { type: string }
    AssignTenantInput:
      type: object
      properties:
        tenantId:
          type: string
          nullable: true
          description: Pass null to unassign
        leaseStart: { type: string, format: date }

    Payment:
      type: object
      required: [id, shopId, shopCode, tenantId, tenantName, month, amountDue, amountPaid, status, createdAt]
      properties:
        id: { type: string }
        shopId: { type: string }
        shopCode: { type: string }
        tenantId: { type: string }
        tenantName: { type: string }
        month: { type: string, description: "YYYY-MM" }
        amountDue: { type: number }
        amountPaid: { type: number }
        status:
          type: string
          enum: [paid, unpaid, partial]
        paidOn: { type: string, format: date }
        method:
          type: string
          enum: [cash, bkash, nagad, rocket, bank, other]
        note: { type: string }
        createdAt: { type: string, format: date-time }
    PaymentInput:
      type: object
      required: [shopId, month, amountDue, amountPaid]
      properties:
        shopId: { type: string }
        month: { type: string }
        amountDue: { type: number }
        amountPaid: { type: number }
        paidOn: { type: string, format: date }
        method:
          type: string
          enum: [cash, bkash, nagad, rocket, bank, other]
        note: { type: string }
    GenerateMonthInput:
      type: object
      required: [month]
      properties:
        month: { type: string, description: "YYYY-MM" }

    DashboardSummary:
      type: object
      required: [totalShops, occupiedShops, vacantShops, totalTenants, currentMonth, expectedThisMonth, collectedThisMonth, outstandingThisMonth, paidCount, unpaidCount]
      properties:
        totalShops: { type: integer }
        occupiedShops: { type: integer }
        vacantShops: { type: integer }
        totalTenants: { type: integer }
        currentMonth: { type: string }
        expectedThisMonth: { type: number }
        collectedThisMonth: { type: number }
        outstandingThisMonth: { type: number }
        paidCount: { type: integer }
        unpaidCount: { type: integer }

    DueReminder:
      type: object
      required: [paymentId, shopCode, tenantName, tenantPhone, month, amountDue, amountPaid, daysOverdue]
      properties:
        paymentId: { type: string }
        shopCode: { type: string }
        tenantName: { type: string }
        tenantPhone: { type: string }
        month: { type: string }
        amountDue: { type: number }
        amountPaid: { type: number }
        daysOverdue: { type: integer }

    MonthlyCollectionPoint:
      type: object
      required: [month, expected, collected]
      properties:
        month: { type: string }
        expected: { type: number }
        collected: { type: number }

    OccupancySummary:
      type: object
      required: [occupied, vacant]
      properties:
        occupied: { type: integer }
        vacant: { type: integer }

    TopTenant:
      type: object
      required: [tenantId, tenantName, totalPaid]
      properties:
        tenantId: { type: string }
        tenantName: { type: string }
        totalPaid: { type: number }