# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shop Rental Management API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalShops: number;
  occupiedShops: number;
  vacantShops: number;
  totalTenants: number;
  currentMonth: string;
  expectedThisMonth: number;
  collectedThisMonth: number;
  outstandingThisMonth: number;
  paidCount: number;
  unpaidCount: number;
}