# lib/api-zod/src/generated/types/dueReminder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shop Rental Management API
 * OpenAPI spec version: 0.1.0
 */

export interface DueReminder {
  paymentId: string;
  shopCode: string;
  tenantName: string;
  tenantPhone: string;
  month: string;
  amountDue: number;
  amountPaid: number;
  daysOverdue: number;
}