# lib/api-zod/src/generated/types/generateMonthInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shop Rental Management API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateMonthInput {
  /** YYYY-MM */
  month: string;
}