# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shop Rental Management API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentMethod } from "./paymentMethod";
import type { PaymentStatus } from "./paymentStatus";

export interface Payment {
  id: string;
  shopId: string;
  shopCode: string;
  tenantId: string;
  tenantName: string;
  /** YYYY-MM */
  month: string;
  amountDue: number;
  amountPaid: number;
  status: PaymentStatus;
  paidOn?: Date;
  method?: PaymentMethod;
  note?: string;
  createdAt: Date;
}